restricted bessel function to integer orders

+++ src/stan/agrad/fwd/bessel_first_kind.hpp
@@ -12,7 +12,7 @@ namespace stan{
     template <typename T>
     inline
     fvar<T>
-    bessel_first_kind(const double v, const fvar<T>& z) {
+    bessel_first_kind(const int v, const fvar<T>& z) {
       using stan::math::bessel_first_kind;
 
       T bessel_first_kind_z(bessel_first_kind(v, z.val_));
